requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
google-cloud-firestore==2.16.1
google-cloud-logging==3.10.0  # optional, but nice
